x86_64/sev: move SEV structures to codebase x86_64/sev: remove SEV detection, replace with a simpler c-bit defining routine
