+++ src/structures/mem_encrypt.rs
@@ -1,0 +1,38 @@
+use core::sync::atomic::{AtomicU64, Ordering};
+
+use crate::structures::paging::page_table::PHYSICAL_ADDRESS_MASK;
+use crate::structures::paging::PageTableFlags;
+
+static C_BIT_MASK: AtomicU64 = AtomicU64::new(0);
+
+/// Enable memory encryption by defining the physical address bit that is used to mark a page
+/// encrypted in a page table entry
+pub fn define_encryption_bit(bit_position: u64) {
+    let c_bit_mask = 1u64 << bit_position;
+
+    PHYSICAL_ADDRESS_MASK.fetch_and(!c_bit_mask, Ordering::AcqRel);
+    C_BIT_MASK.store(c_bit_mask, Ordering::Relaxed);
+}
+
+impl PageTableFlags {
+    #[inline]
+    fn c_bit_mask() -> PageTableFlags {
+        let bit_mask = C_BIT_MASK.load(Ordering::Relaxed);
+        assert_ne!(bit_mask, 0, "C-bit is not set");
+        PageTableFlags::from_bits_retain(bit_mask)
+    }
+
+    /// Sets the encryption bit on the page table entry.
+    ///
+    /// Requires memory encryption to be enabled, or this will panic.
+    pub fn set_encrypted(&mut self, encrypted: bool) {
+        self.set(Self::c_bit_mask(), encrypted);
+    }
+
+    /// Checks if the encryption bit is set on the page table entry.
+    ///
+    /// Requires memory encryption to be enabled, or this will panic.
+    pub fn is_encrypted(&self) -> bool {
+        self.contains(Self::c_bit_mask())
+    }
+}
