+++ src/structures/mod.rs
@@ -6,11 +6,11 @@ pub mod gdt;
 
 pub mod idt;
 
+#[cfg(feature = "amd_sev")]
+pub mod mem_encrypt;
 pub mod paging;
 pub mod port;
 pub mod tss;
-#[cfg(feature = "amd_sev")]
-pub mod amd_sev;
 
 /// A struct describing a pointer to a descriptor table (GDT / IDT).
 /// This is in a format suitable for giving to 'lgdt' or 'lidt'.
